Former-commit-id: 1e6ba6492b58f7811e10df0885a8fa39de7c7e13

+++ src/gameplayer/front_end/gui_generator/button_generator/ButtonFactory.java
@@ -7,7 +7,7 @@ import javafx.scene.input.MouseEvent;
 
 public class ButtonFactory {
 	
-	public Button buildButton(String aMessage, int aXPos, int aYPos, EventHandler<? super MouseEvent> aHandler, ButtonDisplay aDisplayType) {
+	public Button buildButton(String aMessage, double aXPos, double aYPos, EventHandler<? super MouseEvent> aHandler, ButtonDisplay aDisplayType) {
 		IButton buttonClass = null;
 		if (aDisplayType.equals(ButtonDisplay.CSS)) {
 			buttonClass = new CSSButton();
